requirements.txt
discord.py[voice]
yt-dlp
flask
PyNaCl
ffmpeg-python
